examples/browsing_agent/requirements.txt
livekit-agents[openai,silero,turn-detector,deepgram,silero,openai,google,anthropic,cartesia,elevenlabs,rime,noise-cancellation]~=1.0rc
python-dotenv
playwright
Pillow
